feat(socials links): change links for host & artist
